primus-os/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "primus-os"
version = "0.1.0"
description = "AI-Powered Manual Task Automation System"
authors = [
    {name = "Primus OS Team", email = "team@primus-os.com"},
]
dependencies = [
    "openai>=1.0.0",
    "click>=8.0.0",
    "typer>=0.9.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "crewai>=0.1.0",
    "langgraph>=0.0.20",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
primus = "src.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"